Move function `txConstraints` back to `cardano-wallet`.
This function:
- is not needed anywhere in `cardano-balance-tx`;
- is only needed by the wallet's balance migration algorithm.
